+++ listen-kit/Cargo.toml
@@ -7,14 +7,13 @@ description = "Blockchain actions for AI agents"
 documentation = "https://docs.listen-rs.com"
 
 [features]
-default = ["full"]
+default = ["http"]
 full = ["http", "solana", "evm"]
 http = [
   "solana",
   "actix-web",
   "actix-cors",
   "actix-web-lab",
-  "futures-util",
   "tokio-stream",
   "jsonwebtoken",
   "redis",
@@ -56,6 +55,7 @@ futures = "0.3.31"
 serde_with = "3.12.0"
 async-trait = "0.1.85"
 ctor = "0.2.0"
+futures-util = { version = "0.3" }
 
 # evm
 alloy = { version = "0.9", features = ["full"], optional = true }
@@ -77,7 +77,6 @@ spl-associated-token-account = { version = "6.0.0", optional = true }
 actix-web = { version = "4", optional = true }
 actix-cors = { version = "0.6", optional = true }
 actix-web-lab = { version = "0.20", optional = true }
-futures-util = { version = "0.3", optional = true }
 tokio-stream = { version = "0.1.17", optional = true }
 jsonwebtoken = { version = "9.3.0", optional = true }
 redis = { version = "0.28.2", features = ["tokio-comp"], optional = true }
