+++ listen-kit/src/cross_chain/lifi/tokens.rs
@@ -1,6 +1,6 @@
-use alloy::primitives::map::HashMap;
 use serde::{Deserialize, Serialize};
 use serde_json::Number;
+use std::collections::HashMap;
 
 #[derive(Serialize, Deserialize, Debug)]
 #[serde(rename_all = "camelCase")]
