+++ listen-kit/src/evm/mod.rs
@@ -1,4 +1,5 @@
 pub mod abi;
+pub mod agent;
 pub mod balance;
 pub mod data;
 pub mod price;
